Split up web app and connection handling
Extract connection handling into extra class, don't reset server URL on errors, replace toasts with error text.
