+++ app/src/main/res/layout/connect_server.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/connect_server_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -60,6 +61,22 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/connect_title" />
 
+    <TextView
+        android:id="@+id/connection_error_text"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="16dp"
+        android:fontFamily="sans-serif-medium"
+        android:textColor="@color/error_text_color"
+        android:visibility="gone"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/host_input"
+        tools:text="@string/connection_error_cannot_connect"
+        tools:visibility="visible" />
+
     <Button
         android:id="@+id/connect_button"
         style="@style/Widget.AppCompat.Button.Colored"
@@ -70,5 +87,5 @@
         android:textColor="@android:color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/host_input" />
+        app:layout_constraintTop_toBottomOf="@id/connection_error_text" />
 </androidx.constraintlayout.widget.ConstraintLayout>
