+++ app/src/main/res/values-de/strings.xml
@@ -14,7 +14,7 @@
     <string name="connect_to_server_title">Serververbindung herstellen</string>
     <string name="toast_error_cannot_connect_host">Verbindung zu %s konnte nicht hergestellt werden.
 \nPrüfe deine Netzwerkverbindung.</string>
-    <string name="toast_error_invalid_format">Adressformat ungültig</string>
+    <string name="connection_error_invalid_format">Adressformat ungültig</string>
     <string name="toast_error_missing_scheme">Protokoll fehlt (https oder http)</string>
     <string name="pref_music_notification_always_dismissible_title">Musik Player Benachrichtigung immer verwerfbar</string>
     <string name="pref_enable_exoplayer_summary">Aktiviere den experimentellen ExoPlayer Video Player, welcher mehr Videoformate und -Codecs unterstützt und besser ins Betriebssystem integriert ist</string>
