+++ app/src/main/res/values-es-rAR/strings.xml
@@ -12,7 +12,7 @@
     <string name="app_name">Jellyfin</string>
     <string name="toast_error_cannot_connect_host">No se puede establecer la conexión a %s.
 \nCompruebe el nombre de host y su conexión de red.</string>
-    <string name="toast_error_invalid_format">El host tiene un formato no válido</string>
+    <string name="connection_error_invalid_format">El host tiene un formato no válido</string>
     <string name="toast_error_missing_scheme">Esquema faltante (https o http)</string>
     <string name="connect_button_text">Conectar</string>
     <string name="host_input_hint">Host</string>
