+++ app/src/main/res/values-es/strings.xml
@@ -10,7 +10,7 @@
     <string name="battery_optimizations_message">Por favor deshabilita la optimización de la batería para la reproducción de medios mientras la pantalla está apagada.</string>
     <string name="toast_error_cannot_connect_host">No se puede establecer la conexión a %s.
 \nCompruebe el nombre de host y su conexión de red.</string>
-    <string name="toast_error_invalid_format">El host tiene un formato no válido</string>
+    <string name="connection_error_invalid_format">El host tiene un formato no válido</string>
     <string name="toast_error_missing_scheme">Esquema faltante (https o http)</string>
     <string name="connect_button_text">Conectado</string>
     <string name="connect_to_server_title">Conectarse al Servidor</string>
