+++ app/src/main/res/values-fi/strings.xml
@@ -12,7 +12,7 @@
     <string name="pref_music_notification_always_dismissible_title">Aseta musiikkisoittimen ilmoitukset piilotetuiksi</string>
     <string name="pref_category_music_player">Musiikkisoitin</string>
     <string name="activity_name_settings">Jellyfin asetukset</string>
-    <string name="toast_error_invalid_format">Isäntäpalvelimen muoto on virheellinen</string>
+    <string name="connection_error_invalid_format">Isäntäpalvelimen muoto on virheellinen</string>
     <string name="host_input_hint">Isäntäpalvelin</string>
     <string name="toast_error_cannot_connect_host">Yhteyttä %s ei pystytä muodostamaan.
 \nTarkista isäntäpalvelimen nimi ja verkkoyhteytesi.</string>
