+++ app/src/main/res/values-hu/strings.xml
@@ -8,7 +8,7 @@
     <string name="network_title">Engedélyezett hálózat típusok</string>
     <string name="battery_optimizations_message">Ahhoz, hogy a lejátszás kikapcsolt képernyő mellett is működhessen, kapcsold ki az akkumulátor optimalizálást.</string>
     <string name="battery_optimizations_title">Akkumulátor optimalizálás letiltása</string>
-    <string name="toast_error_invalid_format">A hoszt formátuma helytelen</string>
+    <string name="connection_error_invalid_format">A hoszt formátuma helytelen</string>
     <string name="toast_error_missing_scheme">Hiányzó séma (https vagy http)</string>
     <string name="connect_button_text">Kapcsolódás</string>
     <string name="host_input_hint">Hoszt</string>
