+++ app/src/main/res/values-id/strings.xml
@@ -21,7 +21,7 @@
     <string name="battery_optimizations_title">Nonaktifkan Pengoptimalan Baterai</string>
     <string name="toast_error_cannot_connect_host">Koneksi ke %s tidak dapat dibuat.
 \nSilakan periksa nama host dan koneksi jaringan Anda.</string>
-    <string name="toast_error_invalid_format">Host memiliki format yang tidak valid</string>
+    <string name="connection_error_invalid_format">Host memiliki format yang tidak valid</string>
     <string name="downloading">Mengunduh</string>
     <string name="playback_info_and_x_more">… Dan %d lainnya</string>
     <string name="playback_info_audio_streams">Aliran audio</string>
