+++ app/src/main/res/values-ta/strings.xml
@@ -17,7 +17,7 @@
     <string name="battery_optimizations_title">பேட்டரி உகப்பாக்கங்களை முடக்கு</string>
     <string name="toast_error_cannot_connect_host">%s உடன் இணைப்பை நிறுவ முடியாது.
 \nஹோஸ்ட்பெயர் மற்றும் உங்கள் பிணைய இணைப்பை சரிபார்க்கவும்.</string>
-    <string name="toast_error_invalid_format">ஹோஸ்ட் தவறான வடிவத்தைக் கொண்டுள்ளது</string>
+    <string name="connection_error_invalid_format">ஹோஸ்ட் தவறான வடிவத்தைக் கொண்டுள்ளது</string>
     <string name="toast_error_missing_scheme">விடுபட்ட திட்டம் (https அல்லது http)</string>
     <string name="connect_button_text">இணை</string>
     <string name="host_input_hint">ஹோஸ்ட்</string>
